requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
faiss-cpu
pypdf
python-dotenv
transformers
torch
fpdf